pyproject.toml
[tool.poetry]
name = "nwpu-grade"
version = "0.1.0"
description = ""
authors = ["Peng Guanwen <pg999w@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
cssselect = "^1.1.0"
lxml = "^4.6.2"
python-telegram-bot = "^13.1"
pydantic = "^1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
